cppbindings: Move Table into its own source file
